Collect box related functions
