+++ lmnet/lmnet/data_augmentor.py
@@ -22,6 +22,9 @@ import numpy as np
 from PIL import Image, ImageEnhance, ImageFilter
 
 from lmnet import data_processor, pre_processor
+from lmnet.utils.box import fill_dummy_boxes
+from lmnet.utils.box import crop_boxes
+from lmnet.utils.box import iou
 
 
 class Blur(data_processor.Processor):
@@ -381,117 +384,6 @@ class Hue(data_processor.Processor):
         return dict({'image': image}, **kwargs)
 
 
-# TODO(wakisaka): Move to somewhere, It is generic function for object detection.
-def iou(boxes, box):
-    """Calculate overlap
-
-    Args:
-        boxes: boxes in the image. shape is [num_boxes, 4 or more(x, y, w, h, ...)]
-        box: a single box in the image. shape is [4 or more (x, y, w, h, ...)]
-    Returns:
-        iou: shape is [num_boxes]
-    """
-    assert boxes.size != 0, "Cannot calculate if ground truth boxes is zero"
-
-    # format boxes (left, top, right, bottom)
-    boxes = np.stack([
-        boxes[:, 0],
-        boxes[:, 1],
-        boxes[:, 0] + boxes[:, 2],
-        boxes[:, 1] + boxes[:, 3],
-    ], axis=1)
-
-    # format box (left, top, right, bottom)
-    box = np.array([
-        box[0],
-        box[1],
-        box[0] + box[2],
-        box[1] + box[3],
-    ])
-
-    # calculate the left up point
-    left_top = np.maximum(boxes[:, 0:2], box[0:2])
-
-    # calculate the right bottom point
-    right_bottom = np.minimum(boxes[:, 2:], box[2:])
-
-    horizon_vertical = np.maximum(right_bottom - left_top, 0)
-
-    intersection = horizon_vertical[:, 0] * horizon_vertical[:, 1]
-
-    areas = (boxes[:, 2] - boxes[:, 0]) * (boxes[:, 3] - boxes[:, 1])
-    area = (box[2] - box[0]) * (box[3] - box[1])
-
-    epsilon = 1e-10
-    union = area + areas - intersection
-
-    return intersection / (union + epsilon)
-
-
-# TODO(tokunaga): Move to somewhere, It is generic function for object detection.
-def _fill_dummy_boxes(gt_boxes, num_max_boxes):
-    dummy_gt_box = [0, 0, 0, 0, -1]
-    if len(gt_boxes) == 0:
-        gt_boxes = np.array(dummy_gt_box * num_max_boxes)
-        return gt_boxes.reshape([num_max_boxes, 5])
-    elif len(gt_boxes) < num_max_boxes:
-        diff = num_max_boxes - len(gt_boxes)
-        gt_boxes = np.append(gt_boxes, [dummy_gt_box] * diff, axis=0)
-        return gt_boxes
-    return gt_boxes
-
-
-def _crop_boxes(boxes, crop_rect):
-    """Crop boxes
-
-    Args:
-        boxes: ground truth boxes in the image. shape is [num_boxes, 5(x, y, w, h, class)]
-        crop_rect: a crop rectangle. shape is [4(x, y, w, h)]
-    Returns:
-        cropped_boxes: shape is [num_boxes, 5(x, y, w, h, class)]
-    """
-    # check crop_rect overlap with boxes.
-    assert not ((crop_rect[0] + crop_rect[2]) < boxes[:, 0]).any()
-    assert not ((crop_rect[1] + crop_rect[3]) < boxes[:, 1]).any()
-    assert not (crop_rect[0] > (boxes[:, 0] + boxes[:, 2])).any()
-    assert not (crop_rect[1] > (boxes[:, 1] + boxes[:, 3])).any()
-
-    # format to xmin, ymin, xmax, ymax
-    cropped_boxes = np.stack([
-        boxes[:, 0],
-        boxes[:, 1],
-        boxes[:, 0] + boxes[:, 2],
-        boxes[:, 1] + boxes[:, 3],
-    ], axis=1)
-
-    # shift to crop size.
-    cropped_boxes = np.stack([
-        cropped_boxes[:, 0] - crop_rect[0],
-        cropped_boxes[:, 1] - crop_rect[1],
-        cropped_boxes[:, 2] - crop_rect[0],
-        cropped_boxes[:, 3] - crop_rect[1],
-    ], axis=1)
-
-    # adjust beyond box
-    cropped_boxes = np.stack([
-        np.maximum(cropped_boxes[:, 0], 0),
-        np.maximum(cropped_boxes[:, 1], 0),
-        np.minimum(cropped_boxes[:, 2], crop_rect[2]),
-        np.minimum(cropped_boxes[:, 3], crop_rect[3]),
-    ], axis=1)
-
-    # format to x, y, w, h
-    cropped_boxes = np.stack([
-        cropped_boxes[:, 0],
-        cropped_boxes[:, 1],
-        cropped_boxes[:, 2] - cropped_boxes[:, 0],
-        cropped_boxes[:, 3] - cropped_boxes[:, 1],
-        boxes[:, 4],
-    ], axis=1)
-
-    return cropped_boxes
-
-
 class Pad(data_processor.Processor):
     """Add padding to images.
 
@@ -819,8 +711,8 @@ class SSDRandomCrop(data_processor.Processor):
 
                 # take only matching gt boxes
                 masked_boxes = boxes[mask, :]
-                current_boxes = _crop_boxes(masked_boxes, crop_rect)
-                current_boxes = _fill_dummy_boxes(current_boxes, num_max_boxes)
+                current_boxes = crop_boxes(masked_boxes, crop_rect)
+                current_boxes = fill_dummy_boxes(current_boxes, num_max_boxes)
 
                 return dict({'image': current_image, 'gt_boxes': current_boxes}, **kwargs)
